backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlmodel>=0.0.16
SQLAlchemy>=2.0
psycopg[binary]>=3.1
python-jose[cryptography]>=3.3
passlib>=1.7
python-dotenv>=1.0
pydantic>=2.6
email-validator>=2.1
anthropic==0.25.0
python-dateutil==2.8.2
pytest>=8.0
httpx>=0.27
